+++ polypesto/core/problem/__init__.py
@@ -1,14 +1,14 @@
 from pypesto.problem import Problem as PypestoProblem
 
 from .paths import ProblemPaths
-from .base import Problem, write_and_load_problem
+from .base import Problem, write_petab
 from .estimate import run_parameter_estimation
 
 
 __all__ = [
     "PypestoProblem",
     "Problem",
-    "write_and_load_problem",
+    "write_petab",
     "ProblemPaths",
     "run_parameter_estimation",
 ]
