+++ polypesto/core/simulate.py
@@ -1,21 +1,14 @@
-import os
-from dataclasses import dataclass
-from typing import Dict, Generic, List, Optional, Type, TypeVar
+from typing import List
 
-from numpy.typing import ArrayLike
-import pandas as pd
+from amici.petab.simulations import simulate_petab, rdatas_to_measurement_df
+from pypesto.objective import AmiciObjective
 
 from polypesto.core.params import ParameterSet
-from polypesto.core.problem import ProblemPaths, Problem, write_and_load_problem
-from polypesto.models.base import ModelBase
-from polypesto.core.petab import (
-    PetabData,
-    define_empty_measurements,
-    PetabIO,
-    add_noise_to_measurements,
-)
-
-from polypesto.core.experiment import SimConditions
+from polypesto.models import ModelBase
+from polypesto.core.problem import ProblemPaths, Problem, write_petab
+from polypesto.core.pypesto import load_pypesto_problem, PypestoProblem
+import polypesto.core.petab as pet
+from polypesto.core.conditions import SimConditions, define_cond_df
 
 
 def simulate_problem(
@@ -23,11 +16,6 @@ def simulate_problem(
     true_params: ParameterSet,
     **kwargs,
 ) -> Problem:
-
-    from polypesto.core.pypesto import load_pypesto_problem, PypestoProblem
-    from polypesto.core.experiment import create_experiments_from_petab
-    from pypesto.objective import AmiciObjective
-    from amici.petab.simulations import simulate_petab, rdatas_to_measurement_df
 
     importer, pypesto_problem = load_pypesto_problem(
         yaml_path=problem.paths.petab_yaml, model_name=problem.model.name, **kwargs
@@ -52,11 +40,11 @@ def simulate_problem(
         petab_problem.measurement_df,
     )
 
-    meas_df = add_noise_to_measurements(
+    meas_df = pet.add_noise_to_measurements(
         meas_df, noise_level=problem.model.obs_noise_level
     )
 
-    PetabIO.write_meas_df(meas_df, filename=problem.paths.measurements)
+    pet.PetabIO.write_meas_df(meas_df, filename=problem.paths.measurements)
 
     return Problem.load(
         model=problem.model,
@@ -70,8 +58,6 @@ def simulate_experiments(
     conds: List[SimConditions],
 ) -> Problem:
 
-    from polypesto.core.experiment import define_cond_df
-
     cond_df = define_cond_df(conds)
     obs_df = model.get_obs_df()
     param_df = model.get_param_df()
@@ -80,14 +66,15 @@ def simulate_experiments(
         for obs in model.observables.keys():
             obs_id = f"obs_{obs}"
             data_dict[(obs_id, cond.exp_id)] = cond.t_eval
-    meas_df = define_empty_measurements(data_dict)
+    meas_df = pet.define_empty_measurements(data_dict)
 
-    petab_data = PetabData(obs_df, cond_df, param_df, meas_df)
+    petab_data = pet.PetabData(obs_df, cond_df, param_df, meas_df)
     paths = ProblemPaths(data_dir)
 
     true_params = conds[0].true_params
-    problem = write_and_load_problem(model, paths, petab_data, true_params)
+    write_petab(model, paths, petab_data, true_params)
 
+    problem = Problem.load(model, paths)
     problem = simulate_problem(problem, true_params)
 
     return problem
